requirements.txt
alembic==1.13.2
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Mail==0.10.0
Flask-Markdown==0.3
Flask-Migrate==4.0.7
Flask-Misaka==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
misaka==2.1.1
mysqlclient==2.2.4
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
setuptools==75.1.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
wheel==0.44.0
WTForms==3.1.2


